Fallback to the deeplinking configuration for url by default
The behaviour we had in D2L applies to all LMS where we have tried deep
linking being Canvas the exceptions which overrides get_document.

The behaviour is:

- If we have the assignment in the DB use its document
- If not, but we have a record of the original assignment the current
  launch was copied from, use that.

- If not, check if we have deep linking configuration in the launch and
  use that. This is likely the first launch of the assignment.

- Return None otherwise.
