+++ lms/product/d2l/_plugin/misc.py
@@ -12,16 +12,6 @@ class D2LMiscPlugin(MiscPlugin):
         # `lti_registration.token_url` as in other LMS
         return "https://api.brightspace.com/auth/token"
 
-    def get_document_url(self, request, assignment, historical_assignment):
-        url = super().get_document_url(request, assignment, historical_assignment)
-
-        if not url:
-            # In D2L support both deep linking and DB backed assignment.
-            # Use the DL url as a fallback.
-            url = self.get_deep_linked_assignment_configuration(request).get("url")
-
-        return url
-
     @classmethod
     def factory(cls, _context, request):
         return cls()
